Fixed everything

+++ Assignment5/q2_filter_history.py
@@ -8,9 +8,16 @@
 import ctypes
 import sys
 import pickle
+import os
+# Same filtering method as done in Q1
 def filt():
+	# check if the result image exists
+	if os.path.isfile("result.bmp") == False:
+		print "No file loaded! Please load a file."
+		return None
+	# Import the shared library to perform the filter
 	fastFilt = ctypes.cdll.LoadLibrary("./libfast_filter.so")
-
+	#Open the image, get the data and send the data to the undo pickle dump file
 	in_img = open('result.bmp', 'rb')
 	in_img_data = in_img.read()
 	pickle.dump(in_img_data, open("undo.p","wb"))
@@ -30,19 +37,30 @@ def filt():
 
 	out_img_data = " " *len(in_img_data)
 
-	#Perform FIlter
+	#Perform Filter and write image to result.bmp
 	fastFilt.doFiltering(in_img_data, weights, width, out_img_data)
 	out_img = open("result.bmp", "wb")
 	out_img.write(out_img_data)
 	out_img.close()
 def load():
+	#Load in the data from the given bmp file
 	in_img = open(sys.argv[2], 'rb')
 	in_img_data = in_img.read()
 	in_img.close()
  	out_img = open("result.bmp", "wb")
 	out_img.write(in_img_data)
 	out_img.close()
+	# Delete the old pickle dump files if any exist
+	if os.path.isfile("undo.p") == True:
+		os.remove("undo.p")
+	if os.path.isfile("redo.p") == True:
+		os.remove("redo.p")
 def undo():
+	#cheching if undo was done before
+	if os.path.isfile("undo.p") == False:
+		print "Nothing to undo!"
+		return None
+	# load the previous data and undo the previous steps
 	undo_data = pickle.load(open("undo.p","rb"))
 
 	redo = open('result.bmp', 'rb')
@@ -54,6 +72,11 @@ def undo():
 	out_img.write(undo_data)
 	out_img.close()
 def redo():
+	# checking to see if redo is possible
+	if os.path.isfile("redo.p") == False:
+		print "Nothing to redo!"
+		return None
+	# load the data and perform the redo
 	redo_data = pickle.load(open("redo.p","rb"))
 
 	undo = open('result.bmp', 'rb')
@@ -66,7 +89,7 @@ def redo():
 	out_img.close()
 
 
-	
+# checking user input	
 if sys.argv[1] == "load" :
 	load()
 if sys.argv[1] == "filter":
